+++ cloud-agent/service/server/src/main/scala/org/hyperledger/identus/event/controller/http/CreateWebhookNotification.scala
@@ -1,9 +1,13 @@
 package org.hyperledger.identus.event.controller.http
 
+import org.hyperledger.identus.api.http.Annotation
 import sttp.tapir.Schema
+import sttp.tapir.Schema.annotations.{description, encodedExample}
 import zio.json.{DeriveJsonDecoder, DeriveJsonEncoder, JsonEncoder, JsonDecoder}
 
 final case class CreateWebhookNotification(
+    @description(CreateWebhookNotification.annotations.url.description)
+    @encodedExample(CreateWebhookNotification.annotations.url.example)
     url: String,
     customHeaders: Option[Map[String, String]]
 )
@@ -12,4 +16,13 @@ object CreateWebhookNotification {
   given encoder: JsonEncoder[CreateWebhookNotification] = DeriveJsonEncoder.gen[CreateWebhookNotification]
   given decoder: JsonDecoder[CreateWebhookNotification] = DeriveJsonDecoder.gen[CreateWebhookNotification]
   given schema: Schema[CreateWebhookNotification] = Schema.derived
+
+  object annotations {
+    object url
+        extends Annotation[String](
+          description = "A URL of webhook for event notification",
+          example = "http://example.com"
+        )
+  }
+
 }