+++ cloud-agent/service/wallet-api/src/main/scala/org/hyperledger/identus/agent/walletapi/model/KeyManagement.scala
@@ -3,7 +3,7 @@ package org.hyperledger.identus.agent.walletapi.model
 import org.hyperledger.identus.castor.core.model.did.EllipticCurve
 import org.hyperledger.identus.castor.core.model.did.InternalKeyPurpose
 import org.hyperledger.identus.castor.core.model.did.VerificationRelationship
-import io.iohk.atala.prism.crypto.Sha256
+import org.hyperledger.identus.shared.crypto.Sha256Hash
 import org.hyperledger.identus.shared.crypto.DerivationPath
 import org.hyperledger.identus.shared.crypto.Ed25519KeyPair
 import org.hyperledger.identus.shared.crypto.X25519KeyPair
@@ -17,7 +17,7 @@ object WalletSeed {
   extension (s: WalletSeed) {
     final def toString(): String = "<REDACTED>"
     def toByteArray: Array[Byte] = s.toArray
-    def sha256Digest: Array[Byte] = Sha256.compute(toByteArray).getValue()
+    def sha256Digest: Array[Byte] = Sha256Hash.compute(toByteArray).bytes.toArray
   }
 
   def fromByteArray(bytes: Array[Byte]): Either[String, WalletSeed] = {