+++ tests/integration-tests/src/test/kotlin/steps/did/PublishDidSteps.kt
@@ -10,12 +10,12 @@ import io.cucumber.java.en.Then
 import io.cucumber.java.en.When
 import io.iohk.atala.automation.extensions.get
 import io.iohk.atala.automation.serenity.ensure.Ensure
-import org.hyperledger.identus.client.models.*
 import net.serenitybdd.rest.SerenityRest
 import net.serenitybdd.screenplay.Actor
 import org.apache.http.HttpStatus
 import org.apache.http.HttpStatus.SC_CREATED
 import org.apache.http.HttpStatus.SC_OK
+import org.hyperledger.identus.client.models.*
 
 class PublishDidSteps {