+++ tests/integration-tests/src/test/kotlin/steps/proofs/PresentProofSteps.kt
@@ -8,11 +8,11 @@ import io.cucumber.java.en.Then
 import io.cucumber.java.en.When
 import io.iohk.atala.automation.extensions.get
 import io.iohk.atala.automation.serenity.ensure.Ensure
-import org.hyperledger.identus.client.models.*
 import models.PresentationEvent
 import net.serenitybdd.rest.SerenityRest
 import net.serenitybdd.screenplay.Actor
 import org.apache.http.HttpStatus.SC_CREATED
+import org.hyperledger.identus.client.models.*
 
 class PresentProofSteps {